Implement lic_13
This adds an implementation for the lic_13 function that checks whether
both of the following conditions are true:
- any three points with A_PTS and B_PTS consecutive intervening points
cannot all be contained in a circle with radius RADIUS1
- any three points with A_PTS and B_PTS consecutive intervening points
cannot all be contained in a circle with radius RADIUS2
If they both are true, it returns true. Otherwise, it returns false.
It also adds three new unit tests, one for the true case and two for the
false case.

Fixes #16