Implement lic_14
This adds an implementation for the lic_14 function that checks whether
both of the following conditions are true:
- any three points with E_PTS and F_PTS consecutive intervening points
cannot all be contained in a triangle with area AREA1
- any three points with E_PTS and F_PTS consecutive intervening points
can all be contained in a triangle with area AREA2
If they both are true, it returns true. Otherwise, it returns false.
Additionally, this adds four new unit tests.

Fixes #9