Implement lic_8
This adds an implementation for the lic_8 function that checks whether
any three points with A_PTS and B_PTS consecutive intervening points
cannot all be contained in a circle with radius RADIUS1. This also adds
5 unit tests that cover the true case, the false case and edge cases
(i.e. when the triangle is degenerate or there are fewer than 5 points).

Fixes #13