Implement lic_9
This implements lic_9 which checks if there is a set of three points,
separated by [c_pts] and [d_pts] consecutive points, which form an angle
that is smaller than pi - [epsilon], or equivalently, greater than
pi + [epsilon].

Also adds some unit tests to verify the behaviour.

Note that if all points lie on a straight line, the angle will be
considered as 0 and not as pi.

Fixes #4